requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
Babel==2.6.0
beautifulsoup4==4.7.1
Brlapi==0.6.6
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
command-not-found==0.3
comtypes==1.1.7
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
defusedxml==0.5.0
diff-match-patch==20181111
distro-info==0.18
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.38.0
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-classified==0.8.11
django-filter==2.1.0
django-import-export==1.2.0
django-model-utils==3.1.1
django-tables2==2.0.4
et-xmlfile==1.0.1
flake8==3.6.0
fpdf==1.7.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.3.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.0
mccabe==0.6.1
netifaces==0.10.4
numpy==1.15.4
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.0.1
odfpy==1.4.0
olefile==0.45.1
openpyxl==2.5.14
pandas==0.24.2
parsel==1.5.1
pexpect==4.2.1
Pillow==4.0.0
protobuf==3.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyflakes==2.0.0
pygobject==3.26.1
PyHamcrest==1.9.0
pymacaroons==0.13.0
pymongo==3.7.2
PyNaCl==1.1.2
pyOpenSSL==19.0.0
pyparsing==2.3.0
pyperclip==1.7.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
queuelib==1.5.0
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
rsa==4.0
Scrapy==1.6.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.13.2
six==1.11.0
sorl-thumbnail==12.5.0
soupsieve==1.8
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tablib==0.12.1
tqdm==4.31.1
Twisted==18.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
w3lib==1.20.0
wadllib==1.3.2
whitenoise==4.1.2
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
